x509-signature-splice: Support multiple Go versions
Previously, tor-browser-build and ncdns-repro needed to be synchronized
carefully, because tor-browser-build commits to a Go version, while
ncdns-repro commits to an x509-signature-splice version.  This patch
allows the x509-signature-splice project work with arbitrary Go versions
without modifying the ncdns-repro repo, thereby eliminating the need for
careful synchronization of tor-browser-build and ncdns-repro.